requirements.txt
# PRA COREP Reporting Assistant - Dependencies

# LLM API Client
google-genai>=1.0.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Schema Validation
pydantic>=2.0.0

# CLI Output Formatting
tabulate>=0.9.0
streamlit>=1.30.0
